fixed the requested changes
