+++ src/features/accounts/screens/components/signup.styles.ts
@@ -1,0 +1,20 @@
+import { StyleSheet } from "react-native";
+
+export const signupStyles = StyleSheet.create({
+	main: {
+		borderWidth: 2,
+		borderStyle: "solid",
+		borderColor: "black",
+		borderRadius: 10,
+		marginTop: 200,
+	},
+	text: {
+		textAlign: "center",
+		fontSize: 20,
+		fontWeight: "700",
+		margin: 10,
+	},
+	inputBox: {
+		margin: 15,
+	},
+});
