+++ src/features/home/components/home.styles.ts
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+	container: {
+		flex: 1,
+		justifyContent:"center"
+	},
+	font:{
+		color:"red",
+		textAlign:"center",
+		margin:20
+	},
+	button:{
+		marginTop:100
+	}
+});
