+++ src/features/home/screens/Home.screen.tsx
@@ -1,10 +1,9 @@
 // IMPORTS //
 import React from "react";
 import { View, StyleSheet, Text,Button } from "react-native";
-import { createStackNavigator } from '@react-navigation/stack';
-import { NavigationContainer, useNavigation } from "@react-navigation/native";
 
 // STYLES //
+import { styles } from "../components/home.styles";
 
 // COMPONENTS //
 
@@ -16,7 +15,7 @@ import { NavigationContainer, useNavigation } from "@react-navigation/native";
 
 /** Home screen component */
 const HomeScreen: React.FC = () => {
-	const nav=useNavigation();
+	const navigation=useNavigation();
 	// Define Contexts
 
 	// Define States
@@ -31,24 +30,9 @@ const HomeScreen: React.FC = () => {
 	return (
 		<View style={styles.container}>
 			<Text style={styles.font}>Home Screen Works</Text>
-			<Button title="To Next Page" onPress={()=>nav.push("About",{name:"Pratham"})} />
+			<Button title="To Next Page" onPress={()=>navigation.push("About",{name:"Pratham"})} />
 		</View>
 	);
 };
-
-const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		justifyContent:'center'
-	},
-	font:{
-		color:'red',
-		textAlign:'center',
-		margin:20
-	},
-	button:{
-		marginTop:100
-	}
-});
 
 export default HomeScreen;
